requirements.txt
requests==2.29.0
bs4
lxml
psycopg2-binary
selenium